requirements.txt
asgiref==3.2.10
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
Pillow==8.0.0
PyJWT==1.7.1
python-decouple==3.4
pytz==2020.1
sqlparse==0.4.1
